Bites: extract ApplicationContextBase to allow multiple implementations
and move SDL2 and Android code into deriving classes
